+++ src/components/student/Session.tsx
@@ -5,6 +5,7 @@ import {
 } from '@mui/icons-material';
 import { Box, Button, Card, CardHeader } from '@mui/material';
 import { Session } from '../../utils/Session';
+import moment from 'moment';
 
 
 interface SessionProps {
@@ -21,16 +22,16 @@ const SessionCard = ({ session, onCheckIn }: SessionProps) => (
         alignItems: 'center',
         flexWrap: 'wrap',
     }}>
-        <div style={{ flex: '0 0 100px'}}>
-            {session.dayOfWeek}
+        <div style={{ flex: '0 0 150px'}}>
+            { moment(session.start).format('dddd, h:mm a').toString()}
         </div>
          <AccessTimeIcon />
-         <div style={{ flex: '0 0 100px', marginLeft: '10px',}}>
-            {session.timeOfDay}
+         <div style={{ flex: '0 0 150px', marginLeft: '10px',}}>
+            {moment(session.end).format('dddd, h:mm a').toString()}
         </div>
         <PersonIcon/>
         <div style={{ flex: '1 1 auto', marginLeft: '10px',}}>
-            {session.mentor ? session.mentor : <div style={{color:'grey'}}>Not paired with a mentor</div >}
+            {session.mentorID ? session.mentorID : <div style={{color:'grey'}}>Not paired with a mentor</div >}
         </div>
       </div>
     }
