+++ src/utils/Session.ts
@@ -1,9 +1,30 @@
-export interface Session {
-    title:string;
-    dayOfWeek:string;
-    timeOfDay:string;
-    start?: Date;
-    end?: Date;
+import { DateTime } from 'luxon';
+
+  export interface Session {
+    id:string;
+    start:Date;
+    end:Date;
+    libraryID:string;
+    mentorID?:string;
+    studentID?:string;
     meetingLink?: string;
-    mentor?: string;
-}
+  }
+
+  export const Sessions: Session[] =[
+    {
+      id:'0',
+      start: DateTime.local(2022, 4, 28, 8).toJSDate(),
+      end: DateTime.local(2022, 7, 28, 8).toJSDate(),
+      libraryID:'1',
+      mentorID:'1',
+      studentID: '0'
+    },
+    {
+      id:'1',
+      start: DateTime.local(2022, 4, 28, 8).toJSDate(),
+      end: DateTime.local(2022, 7, 28, 8).toJSDate(),
+      libraryID:'1',
+      mentorID:'2',
+      studentID: '4'
+    },
+  ]
